webedi_2023/requirements.txt
amqp==2.5.2
appdirs==1.4.4
appnope==0.1.0
asgiref==3.7.2
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
black==21.5b1
boto3==1.11.15
botocore==1.14.17
celery==4.4.2
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
click==8.0.4
colorama==0.4.3
configparser==4.0.2
coverage==5.0.3
crayons==0.3.0
cryptography==3.3.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.19
django-admin-autocomplete-filter==0.6.1
django-autocomplete-light==3.7.0.dev0
django-better-admin-arrayfield==1.0.5
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.4.0
django-json-widget==1.0.1
django-modelclone==0.7.1
django-multiselectfield==0.1.12
django-redis==4.12.1
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.4.0
docutils==0.15.2
environ==1.0
factory-boy==2.12.0
Faker==4.0.2
furl==2.1.0
future==0.18.2
gunicorn==20.0.4
humanize==2.6.0
idna==2.8
importlib-metadata==1.5.0
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
jmespath==0.9.5
kombu==4.6.8
lazy-object-proxy==1.4.3
mandrill==1.0.59
Markdown==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
mypy-extensions==0.4.3
newrelic==5.12.0.140
num2words==0.5.7
orderedmultidict==1.0.1
packaging==20.3
paramiko==2.7.2
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
python-redis-lock==3.6.0
pytractor-new==0.1.5
pytz==2019.3
redis==3.5.3
regex==2021.4.4
requests==2.22.0
retry==0.9.2
s3transfer==0.3.3
selenium==3.141.0
sentry-sdk==0.14.3
six~=1.12.0
-e git+https://${GITHUB_ACCESS_TOKEN}@github.com/plateiq/spices.git@v219-ece9d4c9#egg=spices[django]
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
webdriver-manager==2.3.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
textract==1.6.3
